#10076: Merge atan_bw to TTNN
