+++ docs/source/ttnn/ttnn/api.rst
@@ -180,6 +180,7 @@ Pointwise Unary
    ttnn/cos_bw
    ttnn/acosh_bw
    ttnn/acos_bw
+   ttnn/atan_bw
    ttnn/sub_bw
    ttnn/frac_bw
    ttnn/trunc_bw
