+++ docs/source/ttnn/ttnn/dependencies/tt_lib.rst
@@ -838,8 +838,6 @@ Backward Operations
 
 .. autofunction:: tt_lib.tensor.polygamma_bw
 
-.. autofunction:: tt_lib.tensor.atan_bw
-
 .. autofunction:: tt_lib.tensor.atanh_bw
 
 .. autofunction:: tt_lib.tensor.asin_bw
