+++ docs/source/ttnn/ttnn/ttnn/atan_bw.rst
@@ -1,0 +1,6 @@
+.. _ttnn.atan_bw:
+
+ttnn.atan_bw
+###############
+
+.. autofunction:: ttnn.atan_bw
