+++ tests/ttnn/unit_tests/operations/backward/test_backward_atan.py
@@ -4,8 +4,8 @@
 
 import torch
 import pytest
-import tt_lib
-from tests.tt_eager.python_api_testing.unit_testing.backward_ops.utility_funcs import (
+import ttnn
+from tests.ttnn.unit_tests.operations.backward.utility_funcs import (
     data_gen_with_range,
     compare_pcc,
 )
@@ -25,7 +25,7 @@ def test_bw_atan(input_shapes, device):
 
     pyt_y = torch.atan(in_data)
 
-    tt_output_tensor_on_device = tt_lib.tensor.atan_bw(grad_tensor, input_tensor)
+    tt_output_tensor_on_device = ttnn.atan_bw(grad_tensor, input_tensor)
 
     in_data.retain_grad()
