+++ tt_eager/tt_dnn/op_library/backward/backward_ops.hpp
@@ -175,11 +175,6 @@ std::vector<Tensor> polygamma_bw(
     int n,
     const MemoryConfig& output_mem_config = operation::DEFAULT_OUTPUT_MEMORY_CONFIG);
 
-std::vector<Tensor> atan_bw(
-    const Tensor& grad,
-    const Tensor& input,
-    const MemoryConfig& output_mem_config = operation::DEFAULT_OUTPUT_MEMORY_CONFIG);
-
 std::vector<Tensor> atanh_bw(
     const Tensor& grad,
     const Tensor& input,
